refactor(nvim): flatten mini module files under lua
